The view is modified and the main logic of the application is updated

+++ data/src/main/java/com/challenge/data/local/dao/ArticleDao.kt
@@ -14,7 +14,7 @@ interface ArticleDao {
     fun getAllArticles(): Flow<List<ArticleEntity>>
 
     @Query("SELECT * FROM articles WHERE id = :articleId")
-    fun getArticleById(articleId: Int): Flow<ArticleEntity>
+    fun getArticleById(articleId: Long): Flow<ArticleEntity>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAll(articles: List<ArticleEntity>)
